Ported mirror example to Python
I ported the mirror example to Python and I believe it is working properly. Still need to do some further testing to make sure.

I have also done a lot of code cleanup and organization and separated the Windows API specific pieces into their own module.

My wrapper class needs to have some work done to it. I did a direct port of the mirror example so it makes all of the raw c-type calls, this is what I want to avoid doing. Python is a simple  programming language and binding to the Windows API using Python is anything but simple. This does show that the port of the Dokany SDK and the bindings to the driver does work properly.