Added and documented constants.
I have added all 1241 Error and Status codes that dokany uses, These have been added to a database of sorts in order to keep the amount of memory used lower. No point in storing something in memory is it is not going to get used. It will be a smidge slower at loading time but nothing that is going to be really noticed. It is a cached system so once a code gets looked up the first time it gets stored, so it is not going to bake repeated pulls from the database file.

The py_dokany.constants module is a dynamic module, what that means is I have replaced the module class instance with an object class instance. I did this so the module will have the ability to call the getattr method in the class.

There is only a single downside to this. The code below is not going to work!!
```python
from py_dokany.constants import ERROR_SUCCESS
```

This will
```python
from py_dokany import constants

ERROR_SUCCESS = constants.ERROR_SUCCESS
```

and also this

```python
from py_dokany import constants

res = son_windows_function_call()
if res == constants.ERROR_SUCCESS:
    # do code here
```

again there is going to me a tine performance decrease only the first time a code gets used, after that it is loaded into memory and there will be no performance hit for later uses of the same code.

The status code or the hresult codes are special, these codes are a wrapper class around `int` so for all intents and purposes it is going to work just like an int will with a single exception. if you print it or wrap it in `str` it will be converted into the description of the code. If you have a need to make a string form of the integer value of it the following will need to be done. `str(int(code))`

There is yet another feature that exists. status and hresult codes can be transformed into an exception. this can be done 2 ways. one of them is using the `constants.make_exception(code)` method the other is using the `code.exception` property.

```python
from py_dokany import constants

raise constants.ERROR_SUCCESS.exception
```

There ar also 2 additional properties to the code, the first one is `name` which returns the string name for the code. The ERROR_SUCCESS code would return `"ERROR_SUCCESS"`. the other is `hex`, this returns the hex value for a code, the value is always going to be filled with leading 0's to represent a 32 bit code.